Update Feature constructor definition to use interface instead of concrete class
Feature defines a constructor with IGeometryObject, Dictionary<string, object>, string types. If an IDictionary<string, object> is provided it will use another Feature constructor defined with IGeometryObject, object, string and will use reflection to create a Dictionary implementation.

When this constructor is used with an IDictionary<string, object> a Runtime Exception is thrown with the message: Parameter count mismatch.

The first constructor should be updated to follow the Liskov Substitution Principle (https://en.wikipedia.org/wiki/Liskov_substitution_principle) to prevent this undesired behavior:

> Substitutability is a principle in object-oriented programming stating that, in a computer program, if S is a subtype of T, then objects of type T may be replaced with objects of type S (i.e. an object of type T may be substituted with any object of a subtype S) without altering any of the desirable properties of the program (correctness, task performed, etc.).

Unit tests were added to verify that the behavior does not change when the interface is used instead of the concrete implementation.

Resolves: #116